pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mllint-example-projects/basic"
version = "0.1.0"
description = "Example project 1: basic"
authors = ["Bart van Oort <bart@vanoort.is>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
pandas = "^1.2.3"
pyaml = "^20.4.0"
scikit-learn = "^0.24.1"
scipy = "^1.6.1"
defusedxml = "^0.7.1"

[tool.poetry.dev-dependencies]
dvc = "^2.0.5"
mllint = "^0.8.5"
pylint = "^2.8.2"
mypy = "^0.812"
black = "^21.5b1"
isort = "^5.8.0"
bandit = "^1.7.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
[[tool.mypy.overrides]]
# Ignore missing type definitions for these packages.
# NOTE: once [this issue](https://github.com/python/mypy/issues/5205) is fixed, 
# then pyproject.toml configuration for mypy will actually work.
# They already updated their documentation it, but didn't actually release the functionality yet.
# TODO: this section works, remove .mypy.ini
module = [
  "defusedxml.*",
  "pandas.*",
  "scipy.*",
  "sklearn.*",
]
ignore_missing_imports = true
